+++ Assassin/BackgroundWorker.h
@@ -9,8 +9,13 @@ License: The MIT License
 
 namespace Assassin
 {
+	using Platform::String;
+	using Platform::StringReference;
 	using Windows::ApplicationModel::Background::IBackgroundTask;
 	using Windows::ApplicationModel::Background::IBackgroundTaskInstance;
+
+	String^ BackgroundWorkerSocketID = 
+		StringReference(L"Assassin.BackgroundWorker");
 	
 	[Windows::Foundation::Metadata::WebHostHidden]
 	public ref class BackgroundWorker sealed : public IBackgroundTask