+++ Assassin/pch.h
@@ -5,3 +5,9 @@
 
 #include "NagisaVersion.h"
 #include "M2AsyncHelpers.h"
+
+#include "OpenSSL\crypto.h"
+#include "OpenSSL\bio.h"
+
+#pragma comment(lib, "libeay32.lib")
+#pragma comment(lib, "ssleay32.lib")